+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * st-input: Export min and max validators to can be validated
 * st-form-directives: Public directive module in order to can be used from outside
 * st-click-outside: Public directive module in order to can be used from outside
+* st-search: Fixed EvenWindowManager elementRef injection
 
 **Others:**